move module review screen
