+++ ui/src/main/java/com/jin/ui/review/content/ReviewHeader.kt
@@ -1,4 +1,4 @@
-package com.jin.honey.feature.review.ui.content
+package com.jin.ui.review.content
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -16,7 +16,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.sp
-import com.jin.honey.R
+import com.jin.ui.R
 import com.jin.ui.theme.ReviewToolbarDividerColor
 
 @Composable
