+++ ui/src/main/java/com/jin/ui/review/content/ReviewItem.kt
@@ -1,4 +1,4 @@
-package com.jin.honey.feature.review.ui.content
+package com.jin.ui.review.content
 
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
@@ -27,7 +27,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.jin.honey.R
+import com.jin.ui.R
 import com.jin.domain.review.ReviewPreview
 import com.jin.ui.theme.ReviewBoxBorderColor
 import com.jin.ui.theme.ReviewDateTextColor
